Transformation to bird view done
Transformation to bird view done, tested and optimized. In order to get
the lines paralel on the transformated image, the functions
"hough_vertices", "region_of_interest" and "hough_lines" from the first
project were used to get the vertices of the transformation polygon.
After that, the transformation was tested with this values on the images
"test1.jpg", "test2.jpg" and "test3.jpg" getting paralel lines on the
three images, indicating that the transformation is valid and can be
used for further analysis.