Pipeline changed.
Pipeline changed and tested with "project_video.mp4" und
"challenge_video.mp4". After that, the pipeline was still not working
good with the challenge video, so the generation of the binary image
with the lane lines was optimized. Adding a mask to only take the
gradient in the channel L when the value on that pixel of the L channel
is higher than a threshold, because we know that the lane lines have a
high L component. After the change, the pipeline works way better on the
challenge video.